+++ __mocks__/lodash.debounce.mock.ts
@@ -1,0 +1,3 @@
+const debounce = jest.fn((fn) => fn);
+
+export default debounce;
