+++ __mocks__/lodash.isEmpty.mock.ts
@@ -1,0 +1,3 @@
+const isEmpty = jest.fn((arr) => arr.length === 0);
+
+export default isEmpty;
