--- __mocks__/openmrs-esm-error-handling.mock.tsx
@@ -1,5 +1,0 @@
-export function createErrorHandler() {
-  return function errorHandler(err) {
-    console.log(`Received error ${err}`);
-  };
-}
