+++ jest.config.json
@@ -5,11 +5,10 @@
   "moduleNameMapper": {
     "^@carbon/icons-react/es/(.*)$": "@carbon/icons-react/lib/$1",
     "^carbon-components-react/es/(.*)$": "carbon-components-react/lib/$1",
-    "@openmrs/esm-api": "<rootDir>/__mocks__/openmrs-esm-api.mock.tsx",
-    "@openmrs/esm-config": "<rootDir>/__mocks__/openmrs-esm-config.mock.tsx",
-    "@openmrs/esm-error-handling": "<rootDir>/__mocks__/openmrs-esm-error-handling.mock.tsx",
-    "@openmrs/esm-react-utils": "<rootDir>/__mocks__/openmrs-esm-react-utils.mock.tsx",
-    "\\.(s?css)$": "identity-obj-proxy"
+    "@openmrs/esm-framework": "<rootDir>/__mocks__/openmrs-esm-framework.mock.tsx",
+    "\\.(s?css)$": "identity-obj-proxy",
+    "lodash-es/debounce": "<rootDir>/__mocks__/lodash.debounce.mock.ts",
+    "lodash-es/isEmpty": "<rootDir>/__mocks__/lodash.isEmpty.mock.ts"
   },
   "setupFiles": ["<rootDir>/src/setup-tests.js"]
 }
